requirements.txt
aiogram==2.25.1
aiohttp==3.8.4
aiosignal==1.3.1
APScheduler==3.10.1
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
croniter==1.3.14
frozenlist==1.3.3
idna==3.4
magic-filter==1.0.9
multidict==6.0.4
pip==23.1.1
pydantic==1.10.7
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
setuptools==47.1.0
six==1.16.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
yarl==1.9.1